axiom-x/Fonterra/backend/requirements.txt
# NZ Drought Early Warning Dashboard - Backend Requirements
# Constitutional AI implementation with Yama principles enforcement

# Core FastAPI framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database and ORM
sqlalchemy==2.0.23
alembic==1.12.1
# psycopg2-binary==2.9.9  # For PostgreSQL (production) - install separately if needed
aiosqlite==0.19.0  # For async SQLite support

# Data processing and scientific computing
numpy==1.24.3
pandas==2.1.3
scipy==1.11.3
matplotlib==3.8.0

# HTTP client for external APIs
httpx==0.25.2
aiohttp==3.9.1

# Caching (Brahmacharya principle - efficient resource use)
redis==5.0.1

# Background job scheduling
apscheduler==3.10.4

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Testing and development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx  # For API testing

# Logging and monitoring
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Data validation and quality
great-expectations==0.17.23

# Geographic utilities
geopy==2.3.0
shapely==2.0.1

# Time zone handling
pytz==2023.3
python-dateutil==2.8.2

# Security and rate limiting
slowapi==0.1.9

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1